requirements.txt
altair==5.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
idna==3.8
importlib-resources==6.4.4
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.5.5
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pandas-datareader==0.10.0
peewee==3.17.6
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.2
plotly==5.23.0
protobuf==5.27.3
pyarrow==17.0.0
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
webencodings==0.5.1
yfinance==0.2.43
zipp==3.20.0
